MLBasedSketchGenProject/requirements.txt
# 核心ML库
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
diffusers>=0.27.2
accelerate>=0.21.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# CLIP相关
open_clip_torch>=2.20.0
clip-anytorch>=2.5.0

# 工具库
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
ipython>=8.14.0
jupyter>=1.0.0

# 用于笔画分析
scikit-learn>=1.3.0
networkx>=3.1

# Markdown报告生成
tabulate>=0.9.0  # 用于表格格式化